[Refactor] Removed project.types.ts and updated tag-edits.tsx and tag-context.ts imports
