+++ apps/web/package.json
@@ -32,7 +32,6 @@
 		"@radix-ui/react-tabs": "^1.1.1",
 		"@radix-ui/react-toast": "^1.2.6",
 		"@radix-ui/react-tooltip": "^1.1.4",
-		"@services/supabase": "workspace:*",
 		"@simplewebauthn/browser": "^13.1.0",
 		"@simplewebauthn/server": "^13.1.1",
 		"@stellar/stellar-sdk": "13.1.0",
